+++ src/lib.js
@@ -3,6 +3,7 @@ import Widget from './widget/index.js';
 import * as views from './views/index.js';
 import * as pages from './pages/index.js';
 import inspector from './inspector/index.js';
+import upload from './core/upload.js';
 import router from './core/router.js';
 import utils from './core/utils/index.js';
 import { buttons as navButtons } from './nav/index.js';
@@ -14,6 +15,7 @@ export {
     pages,
     navButtons,
     inspector,
+    upload,
     router,
     utils
 };